+++ node/clients/index.ts
@@ -8,6 +8,7 @@ import OMSClient from './Oms'
 import StorefrontPermissions from './storefrontPermissions'
 import IdentityClient from './IdentityClient'
 import Catalog from './catalog'
+import SellersClient from './sellers'
 
 // Extend the default IOClients implementation with our own custom clients.
 export class Clients extends IOClients {
@@ -41,5 +42,9 @@ export class Clients extends IOClients {
 
   public get catalog() {
     return this.getOrSet('catalog', Catalog)
+  }
+
+  public get sellers() {
+    return this.getOrSet('sellers', SellersClient)
   }
 }